// src/app/Models/GetAbsDetailByListEmpIDDataClass.ts
import { DayHourMinuteClass } from "./DayHourMinuteClass";


// Generated by https://quicktype.io

export interface GetAbsDetailByListEmpIDDataClass {
    AbsID:         number;
    EmpID:         string;
    DateB:         string;
    TimeB:         string;
    TimeE:         string;
    DateTimeB:     string;
    DateTimeE:     string;
    HoliDayID:     number;
    Use:           number;
    UseDayHourMinute: DayHourMinuteClass;
    SalaryYYMM:    string;
    KeyName:       string;
    EventDate:     string;
    HoliDay:       HoliDay;
    Today:         boolean;
    State:         string;
    RoteRestList:  string;
    Base:          Base;
    DeptCode:      string;
    DeptNameC:     string;
    DeptNameE:     string;
    AttHrs:        number;
    ProcessFlowID: number;
    CreatDate:     string;
}

export interface Base {
    EmpID:    string;
    EmpCode:  string;
    EmpNameC: string;
    EmpNameE: string;
}

export enum EmpNameC {
    張O倩 = "張O倩",
}

export interface HoliDay {
    HoliDayID:        number;
    HoliDayCode:      string;
    HoliDayNameC:     HoliDayNameC;
    HoliDayUnit:      string;
    HoliDayUnitName:  HoliDayUnitName;
    MinNum:           number;
    IncludeHoliDay:   boolean;
    MaxNum:           number;
    CheckRestHour:    boolean;
    Sex:              string;
    IsDisplay:        boolean;
    AbsentUnit:       number;
    HoliDayKindID:    number;
    HoliDayFlag:      HoliDayFlag;
    Seq:              number;
    AutoCreate:       boolean;
    AutoFlowStart:    boolean;
    AbsentLeave:      boolean;
    Memo:             any;
    IsNotRefRote:     boolean;
    IsDisMinus:       boolean;
    IsDisBalance:     boolean;
    IsDisWithHolding: boolean;
}

export enum HoliDayFlag {
    Empty = "-",
}

export enum HoliDayNameC {
    Empty = "****",
}

export enum HoliDayUnitName {
    分鐘 = "分鐘",
    小時 = "小時",
}